refactor: split data sources and expand tests
