+++ tests/test_reporting_realised_apy.py
@@ -2,9 +2,10 @@ from pathlib import Path
 
 import pandas as pd
 
-from stable_yield_lab import HistoricalCSVSource, Pipeline
+from stable_yield_lab import Pipeline
 from stable_yield_lab.core import Pool, PoolRepository
 from stable_yield_lab.reporting import cross_section_report
+from stable_yield_lab.sources import HistoricalCSVSource
 
 
 def test_cross_section_report_includes_realised_apy(tmp_path: Path) -> None:
